new/requirements.txt
APScheduler==3.7.0
boto3==1.18.18
botocore==1.21.18
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cognitojwt==1.4.1
cryptography==3.4.7
ecdsa==0.17.0
Faker==8.11.0
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
marshmallow==3.13.0
mongomock==3.23.0
pyasn1==0.4.8
pycparser==2.20
pymongo==3.12.0
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
sentinels==1.0.0
six==1.16.0
text-unidecode==1.3
tzlocal==2.1
urllib3==1.26.6
Werkzeug==2.0.1
